+++ src/main/java/me/night0721/lilase/events/SniperFlipperEvents.java
@@ -27,6 +27,7 @@ import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
 import net.minecraftforge.fml.common.gameevent.InputEvent;
 import org.lwjgl.input.Keyboard;
 
+import java.awt.*;
 import java.util.ArrayList;
 import java.util.Calendar;
 import java.util.List;
@@ -36,7 +37,7 @@ import java.util.regex.Pattern;
 
 import static me.night0721.lilase.config.AHConfig.GUI_COLOR;
 import static me.night0721.lilase.config.AHConfig.SEND_MESSAGE;
-import static me.night0721.lilase.features.flipper.Flipper.rotation;
+import static me.night0721.lilase.features.flipper.Flipper.*;
 import static me.night0721.lilase.features.flipper.FlipperState.START;
 import static me.night0721.lilase.utils.PlayerUtils.sendPacketWithoutEvent;
 
@@ -44,21 +45,39 @@ public class SniperFlipperEvents {
     private int windowId = 1, price;
     private boolean buying = false, bought = false;
     private final Clock clock = new Clock();
-    private final Pattern
-            AUCTION_SOLD_PATTERN = Pattern.compile("^(.*?) bought (.*?) for ([\\d,]+) coins CLICK$"),
-            BOUGHT_PATTERN = Pattern.compile("You purchased (\\w+(?:\\s+\\w+)*) for ([\\d,]+)\\s*(\\w+)!"),
-            BOUGHT_PATTERN_2 = Pattern.compile("You claimed (.+?) from (.+?)'s auction!"),
-            BOUGHT_PATTERN_3 = Pattern.compile("You (purchased|claimed)( (\\\\d+x))? ([^\\\\s]+(\\\\s+[^\\\\d,]+)*)((,| for) (\\\\d+,?)+ coins?(!)?)?");
-    // TODO: Split into one pattern
+    private final Pattern AUCTION_SOLD_PATTERN = Pattern.compile("^(.*?) bought (.*?) for ([\\d,]+) coins CLICK$"),
+            pattern = Pattern.compile("You purchased (\\w+(?:\\s+\\w+)*) for ([\\d,]+)\\s*(\\w+)|You claimed (.+?) from (.+?)'s auction!|(You (purchased|claimed)( (\\\\d+x))? ([^\\\\s]+(\\\\s+[^\\\\d,]+)*)((,| for) (\\\\d+,?)+ coins?(!)?))?");
     public static final List<String> postedNames = new ArrayList<>();
 
     @SubscribeEvent
     public void onChat(ClientChatReceivedEvent event) throws InterruptedException {
         String message = event.message.getUnformattedText();
-        Matcher matcher = AUCTION_SOLD_PATTERN.matcher(message);
-        Matcher boughtMatcher = BOUGHT_PATTERN.matcher(message);
-        Matcher boughtMatcher2 = BOUGHT_PATTERN_2.matcher(message);
-        Matcher boughtMatcher3 = BOUGHT_PATTERN_3.matcher(message);
+        if (message.startsWith("§6[Auction]")) {
+            Matcher matcher = AUCTION_SOLD_PATTERN.matcher(ScoreboardUtils.cleanSB(message));
+            if (matcher.matches()) {
+                try {
+                    webhook.setUsername("Lilase");
+                    webhook.setAvatarUrl(icon);
+                    webhook.addEmbed(new DiscordWebhook.EmbedObject()
+                            .setTitle("Someone bought an item!")
+                            .setFooter("Purse: " + format.format(Utils.getPurse()), icon)
+                            .addField("Item:", matcher.group(2), true)
+                            .addField("Price:", matcher.group(3), true)
+                            .addField("Purchaser:", matcher.group(1).split("[Auction] ")[1], true)
+                            .setColor(Color.decode("#003153")));
+                    if (SEND_MESSAGE) webhook.execute();
+                    Utils.debugLog("Notified Webhook");
+                } catch (Exception e) {
+                    e.printStackTrace();
+                    Utils.debugLog("Failed to send webhook");
+                }
+                Lilase.sniper.incrementAuctionsFlipped();
+            }
+        }
+        /*
+        §6[Auction] §aphiinix_ §ebought §fImplosion Belt §efor §6900,000 coins §lCLICK
+         */
+        Matcher boughtMatcher = pattern.matcher(message);
         if (!message.contains(":")) {
             if (message.equals("You didn't participate in this auction!")) {
                 Utils.debugLog("Failed to buy item, not fast enough. Closing the menu");
@@ -66,22 +85,12 @@ public class SniperFlipperEvents {
             } else if (message.equals("You don't have enough coins to afford this bid!")) {
                 Utils.debugLog("Failed to buy item, not enough money. Closing the menu");
                 InventoryUtils.clickOpenContainerSlot(49);
-            } else if (message.contains("Your new API key is")) {
-                Utils.debugLog("Detected new API key, saving it to config");
-                Utils.debugLog("Saved new API key to config");
-                String apiKey = message.replace("Your new API key is ", "");
-                Lilase.configHandler.setString("APIKey", apiKey);
-            } else if ((boughtMatcher.matches() || boughtMatcher2.matches() || boughtMatcher3.matches()) && bought) {
+            } else if (boughtMatcher.matches() && bought) {
                 new Thread(() -> {
                     bought = false;
                     Utils.debugLog("Bought an item, starting to sell");
-                    try {
-                        if (SEND_MESSAGE) Lilase.sniper.webhook.execute();
-                    } catch (Exception e) {
-                        Utils.debugLog("Failed to send webhook");
-                    }
-                    Utils.debugLog("Target price " + Cofl.price);
-                    price = Cofl.price;
+                    Utils.debugLog("Target price " + Cofl.target);
+                    price = Cofl.target;
                     try {
                         Thread.sleep(1500);
                     } catch (InterruptedException e) {
@@ -93,26 +102,24 @@ public class SniperFlipperEvents {
                 InventoryUtils.clickOpenContainerSlot(13);
                 Lilase.mc.thePlayer.closeScreen();
                 Utils.debugLog("Cannot post item as the cost is too low, stopping fliiper and starting sniper");
-                Lilase.sniper.toggleAuction();
+                Lilase.cofl.toggleAuction();
                 Flipper.state = FlipperState.NONE;
             } else if (message.contains("Can't create a BIN auction for this item for a PRICE this LOW!")) {
                 Lilase.mc.thePlayer.closeScreen();
                 Utils.debugLog("Cannot post item as the cost is too low, stopping fliiper and starting sniper");
-                Lilase.sniper.toggleAuction();
+                Lilase.cofl.toggleAuction();
                 Flipper.state = FlipperState.NONE;
             } else if (message.contains("You were spawned in Limbo")) {
                 Utils.debugLog("Detected in Limbo, stopping everything for 5 minutes");
                 Utils.addTitle("You got sent to Limbo!");
                 Flipper.state = FlipperState.NONE;
-                if (Lilase.sniper.getOpen()) Lilase.sniper.toggleAuction();
+                if (Lilase.cofl.getOpen()) Lilase.cofl.toggleAuction();
                 Thread.sleep(5000);
                 Utils.sendServerMessage("/lobby");
                 Thread.sleep(5000);
                 Utils.sendServerMessage("/skyblock");
                 Thread bzchillingthread = new Thread(bazaarChilling);
                 bzchillingthread.start();
-            } else if (matcher.matches() && postedNames.contains(matcher.group(2))) {
-                Lilase.sniper.incrementAuctionsFlipped();
             }
         }
     }
@@ -156,7 +163,7 @@ public class SniperFlipperEvents {
             }
             Lilase.mc.thePlayer.sendChatMessage("/hub");
             Thread.sleep(6000);
-            Lilase.sniper.toggleAuction();
+            Lilase.cofl.toggleAuction();
         } catch (Exception ignore) {
         }
 
@@ -175,6 +182,7 @@ public class SniperFlipperEvents {
             if (Flipper.state != FlipperState.NONE) {
                 new Thread(() -> {
                     Utils.debugLog("Interrupting Flipper selling");
+                    Flipper.sendInterrupt();
                     Lilase.mc.thePlayer.closeScreen();
                     Flipper.state = FlipperState.NONE;
                     Lilase.cofl.toggleAuction();
@@ -184,8 +192,8 @@ public class SniperFlipperEvents {
                     Utils.debugLog("[PageFlipper] Interrupting PageFlipper sniping");
                     Lilase.pageFlipper.toggleAuction();
                 }).start();
-            } else if (Lilase.sniper.getOpen()) {
-                Lilase.sniper.toggleAuction();
+            } else if (Lilase.cofl.getOpen()) {
+                Lilase.cofl.toggleAuction();
             }
         }
     }
@@ -214,7 +222,7 @@ public class SniperFlipperEvents {
     @SubscribeEvent(priority = EventPriority.HIGHEST)
     public void onInventoryRendering(GuiScreenEvent.BackgroundDrawnEvent event) {
         String windowName = InventoryUtils.getInventoryName();
-        if (AHConfig.BED_SPAM && (Lilase.sniper.getOpen() || Lilase.pageFlipper.getOpen() || Lilase.cofl.getOpen())) {
+        if (AHConfig.BED_SPAM && (Lilase.pageFlipper.getOpen() || Lilase.cofl.getOpen())) {
             if ("BIN Auction View".equals(windowName)) {
                 ItemStack is = InventoryUtils.getStackInOpenContainerSlot(31);
                 if (is != null) {
@@ -237,17 +245,13 @@ public class SniperFlipperEvents {
                         buying = false;
                         Lilase.mc.thePlayer.closeScreen();
                     }
-                } else {
-                    System.out.println("No item in slot 31, ??");
-//                    buying = false;
-//                    Lilase.mc.thePlayer.closeScreen();
                 }
             }
             if (buying && "Confirm Purchase".equals(windowName)) {
                 System.out.println("Window ID: " + Lilase.mc.thePlayer.openContainer.windowId);
                 Lilase.mc.playerController.windowClick(windowId + 1, 11, 0, 0, Lilase.mc.thePlayer);
                 buying = false;
-                if (Lilase.sniper.buying || Lilase.cofl.getQueue().isRunning()) bought = true;
+                if (Lilase.cofl.getQueue().isRunning()) bought = true;
             }
         }
     }