+++ src/main/java/me/night0721/lilase/features/cofl/QueueItem.java
@@ -7,17 +7,19 @@ import me.night0721.lilase.utils.Utils;
 public class QueueItem {
     private final String command, name;
     private final int price;
+    private final int target;
     public static Flipper flipper = null;
 
-    public QueueItem(String command, String name, int price) {
+    public QueueItem(String command, String name, int price, int target) {
         this.command = command;
         this.name = name;
         this.price = price;
+        this.target = target;
     }
 
     public void openAuction() {
         Utils.debugLog("Executing: " + command);
         Utils.sendServerMessage("/viewauction " + command);
-        flipper = new Flipper(name, price);
+        flipper = new Flipper(name, price, target);
     }
 }