+++ src/main/java/me/night0721/lilase/utils/Utils.java
@@ -7,13 +7,17 @@ import net.minecraft.util.ChatComponentText;
 import net.minecraft.util.EnumChatFormatting;
 import net.minecraft.util.IChatComponent;
 
+import java.util.List;
+import java.util.regex.Matcher;
 import java.util.regex.Pattern;
+import java.util.stream.Collectors;
 
 import static me.night0721.lilase.Lilase.mc;
 
 public class Utils {
     public static IChatComponent header = null, footer = null;
     private static final Pattern PATTERN_HUB_LOCATIONS = Pattern.compile("(forest|village|farm|mountain|wilderness|community|graveyard|bazaar|auction)");
+    private static final Pattern PATTERN_PURSE = Pattern.compile("(Purse|Piggy): (?:§.)?([0-9.,]+)");
     public static EffectState cookie;
 
     public static String translateAlternateColorCodes(String text) {
@@ -48,12 +52,35 @@ public class Utils {
         }
     }
 
+    public static int getPurse() {
+        String purse = "";
+        List<String> matches = ScoreboardUtils.getSidebarLines().stream()
+                .map(ScoreboardUtils::cleanSB)
+                .map(PATTERN_PURSE::matcher)
+                .filter(Matcher::find)
+                .map(Matcher::group)
+                .collect(Collectors.toList());
+        String purseline = matches.get(0);
+        Matcher matcher = PATTERN_PURSE.matcher(purseline);
+        if (matcher.find()) {
+            System.out.println("Group: " + matcher.group());
+            purse = matcher.group(2);
+            purse = purse.replace(",", "");
+            purse = purse.replaceAll("\\..*", "");
+            Utils.debugLog("Purse: " + purse);
+            return Integer.parseInt(purse);
+        }
+        return Integer.parseInt(purse);
+
+    }
+
+
     public static void sendMessage(String message) {
-        mc.thePlayer.addChatMessage(new ChatComponentText(EnumChatFormatting.LIGHT_PURPLE + "" + EnumChatFormatting.BOLD + "[Lilase] " + EnumChatFormatting.RESET + EnumChatFormatting.GREEN + EnumChatFormatting.BOLD + message));
+        mc.thePlayer.addChatMessage(new ChatComponentText(EnumChatFormatting.LIGHT_PURPLE + String.valueOf(EnumChatFormatting.BOLD) + "[Lilase] " + EnumChatFormatting.RESET + EnumChatFormatting.GREEN + EnumChatFormatting.BOLD + message));
     }
 
     public static void debugLog(String message) {
-        mc.thePlayer.addChatMessage(new ChatComponentText(EnumChatFormatting.LIGHT_PURPLE + "" + "[Lilase] " + EnumChatFormatting.RESET + EnumChatFormatting.WHITE + message));
+        mc.thePlayer.addChatMessage(new ChatComponentText(EnumChatFormatting.LIGHT_PURPLE + "[Lilase] " + EnumChatFormatting.RESET + EnumChatFormatting.WHITE + message));
     }
 
     public static void debugLog(String... messages) {